fix: dungeon floor condition not working
