+++ src/main/kotlin/me/owdding/catharsis/features/blocks/BlockReplacements.kt
@@ -22,8 +22,11 @@ import net.minecraft.resources.Identifier
 import net.minecraft.server.packs.resources.ResourceManager
 import net.minecraft.world.level.block.Block
 import net.minecraft.world.level.block.state.BlockState
+import tech.thatgravyboat.skyblockapi.api.area.dungeon.DungeonFloor
 import tech.thatgravyboat.skyblockapi.api.events.base.Subscription
+import tech.thatgravyboat.skyblockapi.api.events.dungeon.DungeonEnterEvent
 import tech.thatgravyboat.skyblockapi.api.events.level.BlockChangeEvent
+import tech.thatgravyboat.skyblockapi.helpers.McClient
 import tech.thatgravyboat.skyblockapi.helpers.McLevel
 import tech.thatgravyboat.skyblockapi.platform.Identifiers
 import tech.thatgravyboat.skyblockapi.utils.extentions.filterValuesNotNull
@@ -52,6 +55,7 @@ object BlockReplacements : PreparingModelLoadingPlugin<Map<Block, LayeredBlockRe
 
     val blocksCache: Cache<BlockPos, BlockState> = CacheBuilder.newBuilder().maximumSize(1000).expireAfterWrite(5.minutes.toJavaDuration()).build<BlockPos, BlockState>()
     private val blockToListen: MutableSet<Block> = mutableSetOf()
+    private val usedDungeonFloors = mutableSetOf<DungeonFloor>()
 
     @JvmStatic
     fun getSound(state: BlockState, pos: BlockPos): BlockSoundDefinition = map[state.block]?.select(state, pos) ?: BlockSoundDefinition.DEFAULT
@@ -153,6 +157,35 @@ object BlockReplacements : PreparingModelLoadingPlugin<Map<Block, LayeredBlockRe
             }
 
             original
+        }
+    }
+
+    fun markAllDirty() {
+        val chunks = McLevel.level.chunkSource.storage.chunks
+        val renderer = McClient.self.levelRenderer
+        for (i in 0 until chunks.length()) {
+            val chunk = McLevel.level.chunkSource.storage.chunks.get(i)
+            if (chunk == null || chunk.isEmpty) continue
+
+            for ((index, section) in chunk.sections.withIndex()) {
+                if (section == null || section.hasOnlyAir()) continue
+                renderer.setSectionDirty(
+                    chunk.pos.x,
+                    chunk.level.getSectionYFromSectionIndex(index),
+                    chunk.pos.z,
+                )
+            }
+        }
+    }
+
+    fun addDungeonFloor(floor: DungeonFloor) {
+        usedDungeonFloors.add(floor)
+    }
+
+    @Subscription
+    fun onFloorUpdate(event: DungeonEnterEvent) {
+        if (this.usedDungeonFloors.contains(event.floor)) {
+            this.markAllDirty()
         }
     }
 }
