+++ src/main/kotlin/me/owdding/catharsis/features/blocks/replacements/conditions/DungeonFloorCondition.kt
@@ -1,6 +1,7 @@
 package me.owdding.catharsis.features.blocks.replacements.conditions
 
 import com.mojang.serialization.MapCodec
+import me.owdding.catharsis.features.blocks.BlockReplacements
 import me.owdding.catharsis.generated.CatharsisCodecs
 import me.owdding.ktcodecs.GenerateCodec
 import net.minecraft.core.BlockPos
@@ -14,6 +15,10 @@ import tech.thatgravyboat.skyblockapi.api.area.dungeon.DungeonFloor
 data class DungeonFloorCondition(
     val floor: DungeonFloor
 ) : BlockCondition {
+    init {
+        BlockReplacements.addDungeonFloor(floor)
+    }
+
     override val codec: MapCodec<out BlockCondition> = CatharsisCodecs.DungeonFloorConditionCodec
 
     override fun check(
